requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
appdirs=1.4.4=pypi_0
braceexpand=0.1.7=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.08.22=h06a4308_0
certifi=2023.7.22=pypi_0
charset-normalizer=3.2.0=pypi_0
click=8.1.7=pypi_0
cmake=3.27.4.1=pypi_0
contourpy=1.1.1=pypi_0
cycler=0.12.1=pypi_0
docker-pycreds=0.4.0=pypi_0
filelock=3.12.4=pypi_0
fonttools=4.43.1=pypi_0
fsspec=2023.9.1=pypi_0
ftfy=6.1.1=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.40=pypi_0
h5py=3.9.0=pypi_0
huggingface-hub=0.17.1=pypi_0
idna=3.4=pypi_0
imagenetv2-pytorch=0.1=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.3.2=pypi_0
kiwisolver=1.4.5=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
lit=16.0.6=pypi_0
littleutils=0.2.2=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.8.0=pypi_0
mpmath=1.3.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.1=pypi_0
numpy=1.26.0=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-cupti-cu11=11.7.101=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.2.10.91=pypi_0
nvidia-cusolver-cu11=11.4.0.1=pypi_0
nvidia-cusparse-cu11=11.7.4.91=pypi_0
nvidia-nccl-cu11=2.14.3=pypi_0
nvidia-nvtx-cu11=11.7.91=pypi_0
ogb=1.3.6=pypi_0
open-clip-torch=2.20.0=pypi_0
openssl=3.0.10=h7f8727e_2
outdated=0.2.2=pypi_0
packaging=23.1=pypi_0
pandas=2.1.0=pypi_0
path=16.7.1=pypi_0
pathtools=0.1.2=pypi_0
pillow=10.0.1=pypi_0
pip=23.2.1=pypi_0
protobuf=3.20.3=pypi_0
psutil=5.9.6=pypi_0
pycocotools=2.0.7=pypi_0
pyparsing=3.1.1=pypi_0
python=3.10.13=h955ad1f_0
python-dateutil=2.8.2=pypi_0
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
regex=2023.8.8=pypi_0
requests=2.31.0=pypi_0
safetensors=0.3.3=pypi_0
scikit-learn=1.3.0=pypi_0
scipy=1.11.2=pypi_0
sentencepiece=0.1.99=pypi_0
sentry-sdk=1.32.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=68.0.0=pypi_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sqlite=3.41.2=h5eee18b_0
sympy=1.12=pypi_0
threadpoolctl=3.2.0=pypi_0
timm=0.9.7=pypi_0
tk=8.6.12=h1ccaba5_0
torch=2.0.1=pypi_0
torchvision=0.15.2=pypi_0
tqdm=4.66.1=pypi_0
triton=2.0.0=pypi_0
typing-extensions=4.7.1=pypi_0
tzdata=2023.3=pypi_0
urllib3=2.0.4=pypi_0
wandb=0.15.12=pypi_0
wcwidth=0.2.6=pypi_0
webdataset=0.2.48=pypi_0
wheel=0.38.4=pypi_0
wilds=2.0.0=pypi_0
xz=5.4.2=h5eee18b_0
zlib=1.2.13=h5eee18b_0
